+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### New features
 * add collect postprocessor
 * add support for access tokens for aws connectors
+* add support for brotli compression
 
 ### Fixes
 * fix reconnect issue in HTTP client